refac: splitting by different modules
Closes #12
