+++ src/bin/delete_user.rs
@@ -1,11 +1,11 @@
 use std::env::args;
-use letsconnectdreams::*;
+use letsconnectdreams::db;
 
 fn main() {
     let username = args()
         .nth(1)
         .expect("Enter a correct username!");
-    let mut conn = establish_connection();
-    let result = delete_user(&mut conn, username);
+    let mut conn = db::establish_connection();
+    let result = db::users::delete_user(&mut conn, username);
     println!("Is deleted? -> {result}");
 }
