+++ src/bin/list_users.rs
@@ -1,8 +1,8 @@
-use letsconnectdreams::{establish_connection, list_users};
+use letsconnectdreams::db;
 
 fn main() {
-    let connection = &mut establish_connection();
-    let users = list_users(connection);
+    let connection = &mut db::establish_connection();
+    let users = db::users::list_users(connection);
 
     for user in users {
         println!("{:#?}", user);
