+++ src/bin/validate_user.rs
@@ -1,8 +1,9 @@
 use std::env::args;
-use letsconnectdreams::*;
+use letsconnectdreams::db;
+use letsconnectdreams::crypto::auth;
 
 fn main() {
-    let mut conn = establish_connection();
+    let mut conn = db::establish_connection();
     let username = args()
         .nth(1)
         .expect("Enter a valid username!");
@@ -10,6 +11,6 @@ fn main() {
         .nth(2)
         .expect("Enter a valid password!");
 
-    let result = validate_password(&mut conn, &username, password).unwrap();
+    let result = auth::validate_password(&mut conn, &username, password).unwrap();
     println!("Is valid? -> {result}");
 }
